+++ common/src/main/java/com/evandev/emixx/feature/stackgroup/data/BannerPatternItemGroup.java
@@ -1,0 +1,18 @@
+package com.evandev.emixx.feature.stackgroup.data;
+
+import dev.emi.emi.api.stack.EmiIngredient;
+import dev.emi.emi.api.stack.EmiStack;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.BannerPatternItem;
+
+public class BannerPatternItemGroup extends StackGroup {
+    public BannerPatternItemGroup() {
+        super(ResourceLocation.withDefaultNamespace("banner_patterns"), null);
+    }
+
+    @Override
+    public boolean match(EmiIngredient stack) {
+        if (!(stack instanceof EmiStack s)) return false;
+        return s.getItemStack().getItem() instanceof BannerPatternItem;
+    }
+}