+++ common/src/main/kotlin/concerrox/emixx/EmiPlusPlusPlatform.kt
@@ -1,9 +1,0 @@
-package concerrox.emixx
-
-import java.nio.file.Path
-
-interface EmiPlusPlusPlatform {
-
-    val configDirectoryPath: Path
-
-}