+++ common/src/main/kotlin/concerrox/emixx/util/EmiDrawContextExtensions.kt
@@ -1,9 +1,0 @@
-package concerrox.emixx.util
-
-import dev.emi.emi.runtime.EmiDrawContext
-
-fun EmiDrawContext.push(action: EmiDrawContext.() -> Unit) = apply {
-    push()
-    action(this)
-    pop()
-}