+++ fabric/src/main/java/com/evandev/emixx/EmiPlusPlusPlatformFabric.java
@@ -1,0 +1,13 @@
+package com.evandev.emixx;
+
+import com.evandev.emixx.platform.IPlatformHelper;
+import net.fabricmc.loader.api.FabricLoader;
+
+import java.nio.file.Path;
+
+public class EmiPlusPlusPlatformFabric implements IPlatformHelper {
+    @Override
+    public Path getConfigDirectory() {
+        return FabricLoader.getInstance().getConfigDir();
+    }
+}