+++ neoforge/src/main/java/com/evandev/emixx/EmiPlusPlusPlatformNeoForge.java
@@ -1,0 +1,13 @@
+package com.evandev.emixx;
+
+import com.evandev.emixx.platform.IPlatformHelper;
+import net.neoforged.fml.loading.FMLPaths;
+
+import java.nio.file.Path;
+
+public class EmiPlusPlusPlatformNeoForge implements IPlatformHelper {
+    @Override
+    public Path getConfigDirectory() {
+        return FMLPaths.CONFIGDIR.get();
+    }
+}